+++ src/docs/asciidoc/api/login/login.adoc
@@ -10,3 +10,16 @@ include::{snippets}/user-oauthLogin/request-fields.adoc[]
 
 include::{snippets}/user-oauthLogin/http-response.adoc[]
 include::{snippets}/user-oauthLogin/response-fields.adoc[]
+
+[[user-save]]
+=== 일반회원 저장
+
+==== HTTP Request
+
+include::{snippets}/user-save/http-request.adoc[]
+include::{snippets}/user-save/request-fields.adoc[]
+
+==== HTTP Response
+
+include::{snippets}/user-save/http-response.adoc[]
+include::{snippets}/user-save/response-fields.adoc[]
