+++ src/docs/asciidoc/api/user/user.adoc
@@ -1,16 +1,3 @@
-[[user-save]]
-=== 일반회원 저장
-
-==== HTTP Request
-
-include::{snippets}/user-save/http-request.adoc[]
-include::{snippets}/user-save/request-fields.adoc[]
-
-==== HTTP Response
-
-include::{snippets}/user-save/http-response.adoc[]
-include::{snippets}/user-save/response-fields.adoc[]
-
 [[user-update]]
 === 회원 초기정보 수정
